requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
freezegun
flake8
Pillow
psycopg2-binary
